Create a workable scatterplot
It's basically a trend-line without the line. It shows a series with the
value as the y-axis and time as the x-axis. Next, we should load another
series in to use as the x-axis and ignore time. It also has no scales,
but this could be alleviated with a simple text svg at each corner with
the coordinates.